auth added

+++ src/graphql/resolvers/orderResolvers.ts
@@ -15,15 +15,25 @@ export const orderResolvers = {
             productId: number;
             quantity: number
         }, context: Context) => {
+            if (!context.userId) throw new Error('Unauthorized');
+            const product = await context.prisma.product.findUnique({
+                where: {id: Number(productId)},
+            })
+            if (!product) throw new Error('product not found');
+            if (product.ownerId === parseInt(context.userId)) throw new Error("User" +
+                " can't buy" +
+                " own" +
+                " product")
             return context.prisma.orderItem.create({
                 data: {
-                    user: { connect: { id: Number(context.userId) } },
+                    user: {connect: {id: Number(context.userId)}},
                     product: {connect: {id: Number(productId)}},
                     quantity,
                 },
             });
         },
         deleteOrderItem: async (parent: any, {id}: { id: number }, context: Context) => {
+            if (!context.userId) throw new Error('Unauthorized');
             return context.prisma.orderItem.delete({where: {id: Number(id)}});
         },
     },
